Refactor: Split to client and server components files
