+++ src/app/(main)/home/layout.tsx
@@ -1,9 +1,10 @@
 import {
-  NavigationBar,
   SideBar,
   MainView,
   NowPlaying,
-} from "@/components/(main)/home/navigation";
+} from "@/app/_components/(main)/home/serverComponents";
+
+import { NavigationBar } from "@/app/_components/(main)/home/serverComponents";
 
 // TODO: login 여부 저장 or 가져오기
 
@@ -15,7 +16,7 @@ export default function HomeLayout({
   return (
     <div className="fixed bg-black p-2 w-full z-40 h-screen">
       <NavigationBar />
-      <SideBar isLogin={false} />
+      <SideBar />
       <MainView>{children}</MainView>
       {/* <NowPlaying /> */}
     </div>
