+++ src/app/_components/(auth)/register/terms/clientComponents.tsx
