+++ src/app/_components/(auth)/register/terms/serverComponents.tsx
