+++ src/app/_hooks/useIsLogin.tsx
@@ -1,0 +1,14 @@
+import { useState } from "react";
+
+export const useIsLogin = () => {
+  const [isLogin, setIsLogin] = useState(false);
+
+  const handleIsLogin = () => {
+    setIsLogin(true);
+  };
+
+  return {
+    isLogin,
+    handleIsLogin,
+  };
+};
