--- src/components/(main)/home/library.tsx
@@ -1,3 +1,0 @@
-export function Library() {
-	return <h1>DummyLibrary</h1>;
-}
