--- src/components/(main)/home/navProfileButtons.tsx
@@ -1,8 +1,0 @@
-export function NavProfileButtons() {
-  return (
-    <div className="">
-      <button>가입하기</button>
-      <button>로그인하기</button>
-    </div>
-  );
-}
