+++ src/stories/(auth)/login/reset-password/email/clientComponents.stories.tsx
@@ -1,4 +1,4 @@
-import { ResetPasswordEmailForm } from "@/components/(auth)/login/reset-password/email/clientComponents";
+import { ResetPasswordEmailForm } from "@/app/_components/(auth)/login/reset-password/email/clientComponents";
 
 export default {
   title: "Components/Login/ResetPassword/Email/clientComponents",
