+++ src/stories/(auth)/register/username/clientComponents.stories.tsx
@@ -1,7 +1,7 @@
 import { JSX, useState } from "react";
 import { RegisterFormProvider } from "@/contexts/registerFormContext"; // Correct import path
 
-import { RegisterUsernameForm } from "@/components/(auth)/register/username/clientComponents";
+import { RegisterUsernameForm } from "@/app/_components/(auth)/register/username/clientComponents";
 
 export default {
   title: "Components/Register/Username/clientComponents", // Updated title for uniqueness
