+++ src/stories/(auth)/register/verification/clientComponents.stories.tsx
@@ -1,4 +1,4 @@
-import { VerificationCodeForm } from "@/components/(auth)/register/verification/clientComponents";
+import { VerificationCodeForm } from "@/app/_components/(auth)/register/verification/clientComponents";
 
 export default {
   title: "Components/Register/Verification/clientComponents", // Updated title for uniqueness
